requirements.txt
aiopg==1.4.0
anyio==3.6.2
arabic-reshaper==3.0.0
asgiref==2.3
asn1crypto==1.5.1
async-timeout==3.0.1
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==2.1.2
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==39.0.1
cssselect2==0.7.0
daphne==2.2.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==0.5.0
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.4.5
django-mama-cas==2.5.0
django-wkhtmltopdf==3.4.0
djangocodemirror==2.1.0
djangorestframework==3.12.2
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
Faker==18.10.1
filelock==3.8.2
google-api-core==1.33.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.0.4
h11==0.14.0
html5lib==1.1
http-ece==1.1.0
httptools==0.5.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
libcst==0.4.9
lxml==4.9.2
MarkupPy==1.14
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.24.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
pandas==2.0.2
phonenumberslite==8.13.0
platformdirs==2.6.0
proto-plus==1.22.1
protobuf==3.20.3
psycopg==3.1.9
psycopg2-binary==2.9.5
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==1.7.1
pymongo==3.12.3
pyOpenSSL==22.1.0
PyPDF2==1.26.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-field==1.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywebpush==1.9.4
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.2
rsa==4.9
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
sqlvalidator==0.0.20
svglib==1.5.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
Twisted==22.10.0
txaio==22.2.1
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.25.11
uvloop==0.17.0
virtualenv==20.17.1
watchgod==0.8.2
webencodings==0.5.1
websockets==10.4
whitenoise==5.2.0
wkhtmltopdf==0.2
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.2
